Add playlist videos when video is from a playlist.
